+++ server/app.ts
@@ -2,14 +2,19 @@ import cors from 'cors';
 import express from 'express';
 import log from 'loglevel';
 import responseTime from 'response-time';
+import organizationsRouterV2 from 'routers/organizationsRouterV2';
 import boundsRouter from './routers/boundsRouter';
 import capturesRouter from './routers/capturesRouter';
+import contractsRouter from './routers/contractsRouter';
 import countriesRouter from './routers/countriesRouter';
+import gisRouter from './routers/gisRouter';
 import growerAccountsRouter from './routers/growerAccountsRouter';
 import organizationsRouter from './routers/organizationsRouter';
 import plantersRouter from './routers/plantersRouter';
 import rawCapturesRouter from './routers/rawCapturesRouter';
 import speciesRouter from './routers/speciesRouter';
+import speciesRouterV2 from './routers/speciesRouterV2';
+import stakeholderRouterV2 from './routers/stakeholderRouterV2';
 import tokensRouter from './routers/tokensRouter';
 import transactionsRouter from './routers/transactionsRouter';
 import treesRouter from './routers/treesRouter';
@@ -59,18 +64,26 @@ app.use(express.json()); // parse application/json
 
 // routers
 app.use('/countries', countriesRouter);
+app.use('/v2/countries', countriesRouter);
 app.use('/trees', treesRouter);
 app.use('/planters', plantersRouter);
 app.use('/organizations', organizationsRouter);
+app.use('/v2/organizations', organizationsRouterV2);
 app.use('/species', speciesRouter);
+app.use('/v2/species', speciesRouterV2);
 app.use('/wallets', walletsRouter);
+app.use('/v2/wallets', walletsRouter);
 app.use('/transactions', transactionsRouter);
 app.use('/tokens', tokensRouter);
+app.use('/v2/tokens', tokensRouter);
 app.use('/v2/captures', capturesRouter);
 app.use('/raw-captures', rawCapturesRouter);
-app.use('/grower-accounts', growerAccountsRouter);
+app.use('/v2/growers', growerAccountsRouter);
 app.use('/v2/trees', treesRouterV2);
 app.use('/bounds', boundsRouter);
+app.use('/gis', gisRouter);
+app.use('/contract', contractsRouter);
+app.use('/v2/stakeholder', stakeholderRouterV2);
 // Global error handler
 app.use(errorHandler);