+++ server/infra/database/CaptureRepository.ts
@@ -13,12 +13,14 @@ export default class CaptureRepository extends BaseRepository<Capture> {
 
   filterWhereBuilder(object, builder) {
     const result = builder;
-    const {
-      whereNulls = [],
-      whereNotNulls = [],
-      whereIns = [],
-      ...parameters
-    } = object;
+    const { ...parameters } = object;
+
+    // parse the nested values, if there
+    const whereNulls = object.whereNulls ? JSON.parse(object.whereNulls) : [];
+    const whereNotNulls = object.whereNotNulls
+      ? JSON.parse(object.whereNotNulls)
+      : [];
+    const whereIns = object.whereIns ? JSON.parse(object.whereIns) : [];
 
     if (parameters.tokenized === 'true') {
       whereNotNulls.push('wallet.token.id');
@@ -28,12 +30,30 @@ export default class CaptureRepository extends BaseRepository<Capture> {
     delete parameters.tokenized;
 
     result.whereNot(`${this.tableName}.status`, 'deleted');
+
     whereNotNulls.forEach((whereNot) => {
-      result.whereNotNull(whereNot);
+      // to map table names to fields for query
+      switch (true) {
+        case whereNot === 'tag_id':
+          result.whereNotNull('treetracker.capture_tag.tag_id');
+          break;
+        default:
+          result.whereNotNull(whereNot);
+      }
+
+      // result.whereNotNull(whereNot);
     });
 
     whereNulls.forEach((whereNull) => {
-      result.whereNull(whereNull);
+      // to map table names to fields for query
+      switch (true) {
+        case whereNull === 'tag_id':
+          result.whereNull('treetracker.capture_tag.tag_id');
+          break;
+        default:
+          result.whereNull(whereNull);
+      }
+      // result.whereNull(whereNull);
     });
 
     whereIns.forEach((whereIn) => {
@@ -55,9 +75,9 @@ export default class CaptureRepository extends BaseRepository<Capture> {
       delete filterObject.endDate;
     }
 
-    if (filterObject.tag) {
-      filterObject[`treetracker.capture_tag.tag_id`] = filterObject.tag;
-      delete filterObject.tag;
+    if (filterObject.tag_id) {
+      filterObject[`treetracker.capture_tag.tag_id`] = filterObject.tag_id;
+      delete filterObject.tag_id;
     }
 
     if (filterObject.id) {
@@ -74,11 +94,18 @@ export default class CaptureRepository extends BaseRepository<Capture> {
       delete filterObject.reference_id;
     }
 
-    // how to get all the parent and children ids for the org -- if they exist?
-    // Stakeholder api? or pass the org ids from the frontend?
+    if (filterObject.grower_reference_id) {
+      result.where(
+        `treetracker.grower_account.reference_id`,
+        '=',
+        filterObject.grower_reference_id,
+      );
+      delete filterObject.grower_reference_id;
+    }
+
     if (filterObject.organization_id) {
       result.where(`${this.tableName}.planting_organization_id`, 'in', [
-        filterObject.organization_id,
+        ...filterObject.organization_id,
       ]);
       delete filterObject.organization_id;
     }
@@ -93,6 +120,15 @@ export default class CaptureRepository extends BaseRepository<Capture> {
     //   );
     //   delete filterObject.organization_ids;
     // }
+
+    if (filterObject.session_id) {
+      result.where(
+        `${this.tableName}.session_id`,
+        '=',
+        filterObject.session_id,
+      );
+      delete filterObject.session_id;
+    }
 
     result.where(filterObject);
   }
@@ -111,6 +147,7 @@ export default class CaptureRepository extends BaseRepository<Capture> {
             t.tags,
             field_data.device_configuration.device_identifier,
             treetracker.grower_account.wallet,
+            treetracker.grower_account.reference_id as grower_reference_id,
             wt.wallet_name,
             wt.token_id,
             tk.id AS wallet_token_id
@@ -133,14 +170,10 @@ export default class CaptureRepository extends BaseRepository<Capture> {
               JOIN wallet.token t ON t.wallet_id = w.id
               JOIN treetracker.grower_account ga ON ga.wallet = w.name
             ) wt ON treetracker.capture.grower_account_id = wt.id
-          ${
-            filter.tag
-              ? `INNER JOIN treetracker.capture_tag
-                  on treetracker.capture_tag.capture_id = treetracker.capture.id
-                 INNER JOIN treetracker.tag
-                  on treetracker.capture_tag.tag_id = treetracker.tag.id`
-              : ''
-          }
+          LEFT JOIN treetracker.capture_tag
+              on treetracker.capture_tag.capture_id = treetracker.capture.id
+          LEFT JOIN treetracker.tag
+              on treetracker.capture_tag.tag_id = treetracker.tag.id
           ${
             filter.tokenized
               ? `LEFT JOIN wallet.wallet
@@ -153,10 +186,11 @@ export default class CaptureRepository extends BaseRepository<Capture> {
         `,
         ),
       )
-      .where((builder) => this.filterWhereBuilder(filter, builder));
+      .where((builder) => this.filterWhereBuilder(filter, builder))
+      .distinct();
 
     promise = promise.orderBy(
-      sort?.order_by || 'created_at',
+      sort?.order_by || 'treetracker.capture.created_at',
       sort?.order || 'desc',
     );
 
@@ -180,8 +214,7 @@ export default class CaptureRepository extends BaseRepository<Capture> {
     const result = await knex
       .select(
         knex.raw(
-          `
-            COUNT(*) AS count
+          `COUNT(*) AS count
           FROM treetracker.capture
           LEFT JOIN (
               SELECT ct.capture_id, array_agg(t.name) AS tags
@@ -199,14 +232,10 @@ export default class CaptureRepository extends BaseRepository<Capture> {
               JOIN wallet.token t ON t.wallet_id = w.id
               JOIN treetracker.grower_account ga ON ga.wallet = w.name
             ) wt ON treetracker.capture.grower_account_id = wt.id
-          ${
-            filter.tag
-              ? `INNER JOIN treetracker.capture_tag
-                  on treetracker.capture_tag.capture_id = treetracker.capture.id
-                 INNER JOIN treetracker.tag
-                  on treetracker.capture_tag.tag_id = treetracker.tag.id`
-              : ''
-          }
+          LEFT JOIN treetracker.capture_tag
+              on treetracker.capture_tag.capture_id = treetracker.capture.id
+          LEFT JOIN treetracker.tag
+              on treetracker.capture_tag.tag_id = treetracker.tag.id
           ${
             filter.tokenized
               ? `LEFT JOIN wallet.wallet
@@ -219,7 +248,8 @@ export default class CaptureRepository extends BaseRepository<Capture> {
         `,
         ),
       )
-      .where((builder) => this.filterWhereBuilder(filter, builder));
+      .where((builder) => this.filterWhereBuilder(filter, builder))
+      .distinct();
 
     return result[0].count;
   }
@@ -237,6 +267,7 @@ export default class CaptureRepository extends BaseRepository<Capture> {
           field_data.device_configuration.device AS device_type,
           field_data.device_configuration.os_version AS device_os_version,
           treetracker.grower_account.wallet,
+          treetracker.grower_account.reference_id as grower_reference_id,
           regions.region.properties AS region_properties
           FROM treetracker.capture
             LEFT JOIN treetracker.grower_account