+++ server/infra/database/WalletsRepository.ts
@@ -9,6 +9,7 @@ type Filter = Partial<{ name: string }>;
 export default class WalletsRepository extends BaseRepository<Wallets> {
   constructor(session: Session) {
     super('wallet.wallet', session);
+    this.tableName = 'wallet.wallet';
   }
 
   async getWalletByIdOrName(walletIdOrName: string) {
@@ -17,7 +18,8 @@ export default class WalletsRepository extends BaseRepository<Wallets> {
       wallet.wallet.id,
       wallet.wallet.name,
       wallet.wallet.logo_url,
-      wallet.wallet.created_at
+      wallet.wallet.created_at,
+      wallet.wallet.about
     FROM
      wallet.wallet
     WHERE
@@ -43,7 +45,7 @@ export default class WalletsRepository extends BaseRepository<Wallets> {
 
   async getWalletTokenContinentCount(walletIdOrName: string) {
     const sql = `
-    select continent.name as continent ,count(continent.name) as token_count
+    select continent.name as continent , wallet.wallet.about, count(continent.name) as token_count
     from wallet.wallet
       left join wallet.token on 
         wallet.token.wallet_id = wallet.wallet.id
@@ -54,7 +56,7 @@ export default class WalletsRepository extends BaseRepository<Wallets> {
           and continent.type_id in (select id from region_type where type = 'continents' )
       where wallet.wallet.id::text  = '${walletIdOrName}'
         or wallet.wallet.name = '${walletIdOrName}' 
-      group by continent.name
+      group by continent.name, wallet.about
   `;
 
     const object = await this.session.getDB().raw(sql);
@@ -74,7 +76,8 @@ export default class WalletsRepository extends BaseRepository<Wallets> {
         wallet.wallet.id,
         wallet.wallet.name,
         wallet.wallet.logo_url,
-        wallet.wallet.created_at
+        wallet.wallet.created_at,
+        wallet.wallet.about
       FROM wallet.wallet
       LIMIT ${limit}
       OFFSET ${offset}
@@ -95,7 +98,8 @@ export default class WalletsRepository extends BaseRepository<Wallets> {
         wallet.wallet.id,
         wallet.wallet.name,
         wallet.wallet.logo_url,
-        wallet.wallet.created_at
+        wallet.wallet.created_at,
+        wallet.wallet.about
       FROM wallet.wallet
       WHERE name LIKE '%${keyword}%'
       ORDER BY name
@@ -117,7 +121,8 @@ export default class WalletsRepository extends BaseRepository<Wallets> {
         wallet.wallet.id as id,
         wallet.wallet.name,
         wallet.wallet.logo_url,
-        wallet.wallet.created_at
+        wallet.wallet.created_at,
+        wallet.wallet.about
       FROM wallet.wallet
       join (
       --- convert json array to row
@@ -133,5 +138,21 @@ export default class WalletsRepository extends BaseRepository<Wallets> {
       this.session,
     );
     return objectPatched;
+  }
+
+  async getCount(filter: Filter) {
+    const knex = this.session.getDB();
+
+    const result = await knex
+      .select(
+        knex.raw(`
+        COUNT(DISTINCT(${this.tableName}.id)) AS count
+        FROM ${this.tableName}
+        ${filter.name ? `WHERE name LIKE '%${filter.name}%'` : ''}
+    `),
+      )
+      .first();
+
+    return result.count;
   }
 }