+++ server/interfaces/GrowerAccountFilter.ts
@@ -7,12 +7,14 @@ interface GrowerAccountFilter extends DbModel {
   limit?: number;
   offset?: number;
   keyword?: string;
-  organization_id?: string;
+  organization_id?: string[] | string;
   person_id?: string;
   device_identifier?: string;
   wallet?: string;
   email?: string;
   phone?: string;
+  captures_amount_min?: number;
+  captures_amount_max?: number;
 }
 
 export default GrowerAccountFilter;