+++ server/models/OrganizationV2.ts
@@ -1,23 +1,32 @@
 import log from 'loglevel';
+import OrganizationRepositoryV2 from 'infra/database/OrganizationRepositoryV2';
 import FilterOptions from 'interfaces/FilterOptions';
 import Organization from 'interfaces/Organization';
 import { delegateRepository } from '../infra/database/delegateRepository';
-import OrganizationRepository from '../infra/database/OrganizationRepository';
 
 type Filter = Partial<{
   planter_id: number;
   organization_id: number;
-  ids: Array<number>;
+  grower_id: string;
+  ids: Array<string>;
 }>;
 
 function getByFilter(
-  organizationRepository: OrganizationRepository,
+  organizationRepository: OrganizationRepositoryV2,
 ): (filter: Filter, options: FilterOptions) => Promise<Organization[]> {
   return async function (filter: Filter, options: FilterOptions) {
     if (filter.planter_id) {
       log.warn('using planter filter...');
       const trees = await organizationRepository.getByPlanter(
         filter.planter_id,
+        options,
+      );
+      return trees;
+    }
+    if (filter.grower_id) {
+      log.warn('using grower filter...');
+      const trees = await organizationRepository.getByGrower(
+        filter.grower_id,
         options,
       );
       return trees;
@@ -42,14 +51,16 @@ function getOrganizationLinks(organization) {
 }
 
 export default {
-  getById: delegateRepository<OrganizationRepository, Organization>('getById'),
-  getByMapName: delegateRepository<OrganizationRepository, Organization>(
+  getById: delegateRepository<OrganizationRepositoryV2, Organization>(
+    'getById',
+  ),
+  getByMapName: delegateRepository<OrganizationRepositoryV2, Organization>(
     'getByMapName',
   ),
   getByFilter,
   getOrganizationLinks,
   getFeaturedOrganizations: delegateRepository<
-    OrganizationRepository,
+    OrganizationRepositoryV2,
     Organization
   >('getFeaturedOrganizations'),
 };