+++ server/routers/growerAccountsRouter.ts
@@ -1,6 +1,6 @@
 import express from 'express';
 import Joi from 'joi';
-import { handlerWrapper } from './utils';
+import { handlerWrapper, queryFormatter } from './utils';
 import GrowerAccountRepository from '../infra/database/GrowerAccountRepository';
 import Session from '../infra/database/Session';
 import GrowerAccountFilter from '../interfaces/GrowerAccountFilter';
@@ -11,14 +11,18 @@ const router = express.Router();
 router.get(
   '/count',
   handlerWrapper(async (req, res) => {
+    const query = queryFormatter(req);
+
+    // verify filter values
     Joi.assert(
-      req.query,
+      query,
       Joi.object().keys({
         limit: Joi.number().integer().min(1).max(1000),
         offset: Joi.number().integer().min(0),
         keyword: Joi.string(),
-        organization_id: Joi.number().integer().min(0),
+        organization_id: Joi.array().items(Joi.string().uuid()),
         id: Joi.string().uuid(),
+        reference_id: Joi.number(),
         person_id: Joi.string().uuid(),
         device_identifier: Joi.string(),
         first_name: Joi.string(),
@@ -26,9 +30,14 @@ router.get(
         wallet: Joi.string(),
         email: Joi.string(),
         phone: Joi.string(),
+        captures_amount_min: Joi.number().integer().min(0),
+        captures_amount_max: Joi.number().integer().min(0),
+        whereNulls: Joi.array(),
+        whereNotNulls: Joi.array(),
+        whereIns: Joi.array(),
       }),
     );
-    const { ...rest } = req.query;
+    const { ...rest } = query;
 
     const repo = new GrowerAccountRepository(new Session());
     const { count } = await GrowerAccountModel.getCount(repo)({ ...rest });
@@ -58,6 +67,41 @@ router.get(
 );
 
 router.get(
+  '/wallets',
+  handlerWrapper(async (req, res) => {
+    // verify filter values
+    Joi.assert(
+      req.params,
+      Joi.object().keys({
+        limit: Joi.number().integer().min(1).max(1000),
+        offset: Joi.number().integer().min(0),
+        wallet: Joi.string(),
+      }),
+    );
+
+    const { limit = 20, offset = 0, ...rest } = req.query;
+
+    const repo = new GrowerAccountRepository(new Session());
+    const filter: GrowerAccountFilter = { ...rest };
+
+    const result = await GrowerAccountModel.getWalletsByFilter(repo)(filter, {
+      limit,
+      offset,
+    });
+
+    const { count } = await GrowerAccountModel.getWalletsCount(repo)(filter);
+
+    res.send({
+      total: Number(count),
+      offset,
+      limit,
+      wallets: result,
+    });
+    res.end();
+  }),
+);
+
+router.get(
   '/:id/selfies',
   handlerWrapper(async (req, res) => {
     Joi.assert(req.params.id, Joi.string().uuid().required());
@@ -76,6 +120,7 @@ router.get(
     const repo = new GrowerAccountRepository(new Session());
     const exe = GrowerAccountModel.getById(repo);
     const result = await exe(req.params.id);
+    result.links = GrowerAccountModel.getGrowerAccountLinks(result);
     res.send(result);
     res.end();
   }),
@@ -84,14 +129,18 @@ router.get(
 router.get(
   '/',
   handlerWrapper(async (req, res) => {
+    const query = queryFormatter(req);
+
+    // verify filter values
     Joi.assert(
-      req.query,
+      query,
       Joi.object().keys({
         limit: Joi.number().integer().min(1).max(1000),
         offset: Joi.number().integer().min(0),
         keyword: Joi.string(),
-        organization_id: Joi.string().uuid(),
+        organization_id: Joi.array().items(Joi.string().uuid()),
         id: Joi.string().uuid(),
+        reference_id: Joi.number(),
         person_id: Joi.string().uuid(),
         device_identifier: Joi.string(),
         first_name: Joi.string(),
@@ -99,9 +148,15 @@ router.get(
         wallet: Joi.string(),
         email: Joi.string(),
         phone: Joi.string(),
+        captures_amount_min: Joi.number().integer().min(0),
+        captures_amount_max: Joi.number().integer().min(0),
+        whereNulls: Joi.array(),
+        whereNotNulls: Joi.array(),
+        whereIns: Joi.array(),
       }),
     );
-    const { limit = 20, offset = 0, keyword, ...rest } = req.query;
+
+    const { limit = 20, offset = 0, keyword, ...rest } = query;
 
     const repo = new GrowerAccountRepository(new Session());
     const filter: GrowerAccountFilter = { ...rest };
@@ -111,6 +166,7 @@ router.get(
         limit,
         offset,
       });
+
       res.send({
         total: null,
         offset,
@@ -125,6 +181,7 @@ router.get(
         limit,
         offset,
       });
+
       const { count } = await GrowerAccountModel.getCount(repo)(filter);
       res.send({
         total: Number(count),
@@ -132,6 +189,7 @@ router.get(
         limit,
         grower_accounts: result.map((planter) => ({
           ...planter,
+          links: GrowerAccountModel.getGrowerAccountLinks(planter),
         })),
       });
       res.end();