+++ server/routers/organizationsRouterV2.ts
@@ -1,15 +1,15 @@
 import express from 'express';
 import Joi from 'joi';
 import { handlerWrapper } from './utils';
-import OrganizationRepository from '../infra/database/OrganizationRepository';
+import OrganizationRepositoryV2 from '../infra/database/OrganizationRepositoryV2';
 import Session from '../infra/database/Session';
-import OrganizationModel from '../models/Organization';
+import OrganizationModel from '../models/OrganizationV2';
 
 type Filter = Partial<{
   planter_id: number;
   organization_id: number;
-  ids: Array<number>;
-  stakeholder_uuid: string;
+  grower_id: string;
+  ids: Array<string>;
 }>;
 
 const router = express.Router();
@@ -17,7 +17,7 @@ const router = express.Router();
 router.get(
   '/featured',
   handlerWrapper(async (req, res) => {
-    const repo = new OrganizationRepository(new Session());
+    const repo = new OrganizationRepositoryV2(new Session());
     const result = await OrganizationModel.getFeaturedOrganizations(repo)();
     res.send({
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -35,13 +35,13 @@ router.get(
   '/:id',
   handlerWrapper(async (req, res) => {
     Joi.assert(req.params.id, Joi.string().required());
-    const repo = new OrganizationRepository(new Session());
-    let exe;
-    if (isNaN(req.params.id)) {
-      exe = OrganizationModel.getByMapName(repo);
-    } else {
-      exe = OrganizationModel.getById(repo);
-    }
+    const repo = new OrganizationRepositoryV2(new Session());
+    const exe = OrganizationModel.getById(repo);
+    // if (isNaN(req.params.id)) {
+    //   exe = OrganizationModel.getByMapName(repo);
+    // } else {
+    //   exe = OrganizationModel.getById(repo);
+    // }
     const result = await exe(req.params.id);
     result.links = OrganizationModel.getOrganizationLinks(result);
     res.send(result);
@@ -58,26 +58,19 @@ router.get(
       query,
       Joi.object().keys({
         planter_id: Joi.number().integer().min(0),
-        ids: Joi.array().items(Joi.number()),
-        stakeholder_uuid: Joi.string().uuid(),
+        grower_id: Joi.string().guid(),
+        ids: Joi.array().items(Joi.string().uuid()),
         limit: Joi.number().integer().min(1).max(1000),
         offset: Joi.number().integer().min(0),
       }),
     );
-    const {
-      limit = 20,
-      offset = 0,
-      planter_id,
-      ids = [],
-      stakeholder_uuid,
-    } = query;
-    const repo = new OrganizationRepository(new Session());
+    const { limit = 20, offset = 0, planter_id, grower_id, ids = [] } = query;
+    const repo = new OrganizationRepositoryV2(new Session());
     const filter: Filter = {};
     if (planter_id) {
       filter.planter_id = planter_id;
-    }
-    if (stakeholder_uuid) {
-      filter.stakeholder_uuid = stakeholder_uuid;
+    } else if (grower_id) {
+      filter.grower_id = grower_id;
     }
     if (ids.length) {
       filter.ids = ids;